+++ webv2/helpers/helpers.go
@@ -40,17 +40,6 @@ func UpdateSessionFile() error {
 	return nil
 }
 
-// ContainString check string is present in given list.
-func ContainString(fc []string, col string) string {
-
-	for _, s := range fc {
-		if s == col {
-			return col
-		}
-	}
-	return ""
-}
-
 // RemoveSchemaIssue removes issue from the given list.
 func RemoveSchemaIssue(schemaissue []internal.SchemaIssue, issue internal.SchemaIssue) []internal.SchemaIssue {
 
@@ -71,40 +60,6 @@ func IsSchemaIssuePresent(schemaissue []internal.SchemaIssue, issue internal.Sch
 		}
 	}
 	return false
-}
-
-// DuplicateInArray checks if there is any duplicate element present in the list.
-func DuplicateInArray(element []int) int {
-	visited := make(map[int]bool, 0)
-	for i := 0; i < len(element); i++ {
-		if visited[element[i]] == true {
-			return element[i]
-		} else {
-			visited[element[i]] = true
-		}
-	}
-	return -1
-}
-
-// Difference gives list of element that are only present in first list.
-func Difference(listone, listtwo []int) []int {
-
-	hashmap := make(map[int]int, len(listtwo))
-
-	for _, val := range listtwo {
-		hashmap[val]++
-	}
-
-	var diff []int
-
-	for _, val := range listone {
-
-		_, found := hashmap[val]
-		if !found {
-			diff = append(diff, val)
-		}
-	}
-	return diff
 }
 
 // RemoveSchemaIssues remove all  Hotspot and Interleaved from given list.
