+++ webv2/primarykey/primarykey_service.go
@@ -20,8 +20,8 @@ package primarykey
 
 import (
 	"github.com/cloudspannerecosystem/harbourbridge/spanner/ddl"
-	helpers "github.com/cloudspannerecosystem/harbourbridge/webv2/helpers"
 	"github.com/cloudspannerecosystem/harbourbridge/webv2/session"
+	"github.com/cloudspannerecosystem/harbourbridge/webv2/utility"
 )
 
 // getColumnId return ColumnId for given columnName.
@@ -98,7 +98,7 @@ func isValidColumnIds(pkRequest PrimaryKeyRequest, spannertable ddl.CreateTable)
 
 	cidRequestList := getColumnIdListFromPrimaryKeyRequest(pkRequest)
 	cidSpannerTableList := getColumnIdListOfSpannerTable(spannertable)
-	leftjoin := helpers.Difference(cidRequestList, cidSpannerTableList)
+	leftjoin := utility.Difference(cidRequestList, cidSpannerTableList)
 
 	if len(leftjoin) > 0 {
 		return false
@@ -115,7 +115,7 @@ func isValidColumnOrder(pkRequest PrimaryKeyRequest) bool {
 		list = append(list, pkRequest.Columns[i].Order)
 	}
 
-	if helpers.DuplicateInArray(list) == -1 {
+	if utility.DuplicateInArray(list) == -1 {
 		return false
 	}
